app, libgimp, pdb: fix “Stack smashing detected” crash.
Building with clang and various stack protection flags (as is done in
our flatpak apparently), we can trigger a crash when a gint on the stack
is set as a pointer to gsize.
Anyway now all array size must be gsize.

See: https://discourse.gnome.org/t/problem-with-select-polygon-in-gimp-2-99/24753

+++ app/pdb/image-select-cmds.c
@@ -290,7 +290,7 @@ image_select_polygon_invoker (GimpProcedure         *procedure,
   gboolean success = TRUE;
   GimpImage *image;
   gint operation;
-  gint num_segs;
+  gsize num_segs;
   const gdouble *segs;
 
   image = g_value_get_object (gimp_value_array_index (args, 0));
