+++ libgimp/gimpimageselect_pdb.h
@@ -63,7 +63,7 @@ gboolean gimp_image_select_ellipse          (GimpImage      *image,
                                              gdouble         height);
 gboolean gimp_image_select_polygon          (GimpImage      *image,
                                              GimpChannelOps  operation,
-                                             gint            num_segs,
+                                             gsize           num_segs,
                                              const gdouble  *segs);
 gboolean gimp_image_select_item             (GimpImage      *image,
                                              GimpChannelOps  operation,
